+++ frontend/pages/admin/TeamManagementPage/TeamDetailsWrapper/MembersPage/components/AddMemberModal/AddMemberModal.tsx
@@ -5,7 +5,7 @@ import { IUser } from "interfaces/user";
 import endpoints from "fleet/endpoints";
 import Modal from "components/Modal";
 import Button from "components/buttons/Button";
-import AutocompleteDropdown from "pages/admin/TeamManagementPage/TeamDetailsWrapper/MembersPagePage/components/AutocompleteDropdown";
+import AutocompleteDropdown from "pages/admin/TeamManagementPage/TeamDetailsWrapper/MembersPage/components/AutocompleteDropdown";
 import { IDropdownOption } from "interfaces/dropdownOption";
 
 const baseClass = "add-member-modal";
